Move DarkModeUtils to Lib module
